backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.23
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
chromadb>=0.4.22
python-multipart>=0.0.6
python-docx>=1.1.0
pypdf>=3.17.0
aiofiles>=23.2.1
python-dotenv>=1.0.0
